# Update: optimize stuff in config loader.