--- a/composer.json
+++ b/composer.json
@@ -18,7 +18,7 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^9.3",
-        "vimeo/psalm": "4.0.1",
+        "vimeo/psalm": "^6.13.1",
         "phpstan/phpstan": "1.9.x-dev",
         "laravel/pint": "1.2.*"
     },