--- a/composer.json
+++ b/composer.json
@@ -14,12 +14,12 @@
         "check": "./vendor/bin/phpstan analyse --level 8 src tests"
     },
     "require": {
-        "php": ">=8.0"
+        "php": ">=8.1"
     },
     "require-dev": {
         "phpunit/phpunit": "^9.3",
         "vimeo/psalm": "^6.13.1",
-        "phpstan/phpstan": "1.9.x-dev",
+        "phpstan/phpstan": "^2.0",
         "laravel/pint": "1.2.*"
     },
     "minimum-stability": "dev"