Tested via -
`vendor/bin/phpunit tests/Config/PerformanceTest.php`
```php
<?php

namespace Utopia\Tests;

use PHPUnit\Framework\TestCase;
use Utopia\Config\Config;

/**
 * Original implementation for comparison (uses array_shift)
 */
class ConfigOriginal
{
    /** @var array<string, mixed> */
    public static array $params = [];

    public static function getParam(string $key, mixed $default = null): mixed
    {
        $key = \explode('.', $key);
        $value = $default;
        $node = self::$params;

        while (! empty($key)) {
            $path = \array_shift($key);
            $value = (isset($node[$path])) ? $node[$path] : $default;

            if (isset($node[$path]) && \is_array($value)) {
                $node = $node[$path];
            }
        }

        return $value;
    }
}

class PerformanceTest extends TestCase
{
    private const N_ITEMS = 10000;

    private const N_ROUNDS = 100;

    private const DEPTH = 10;

    public function test_performance_benchmark(): void
    {
        // Generate test data
        [$config, $flatKeys, $deepKeys] = $this->generateTestData(self::N_ITEMS, self::DEPTH);

        // Benchmark ORIGINAL implementation
        ConfigOriginal::$params = $config;
        $origFlatTime = $this->benchmarkLookupsOriginal($flatKeys, self::N_ROUNDS);
        $origDeepTime = $this->benchmarkLookupsOriginal($deepKeys, self::N_ROUNDS);
        $origTotal = $origFlatTime + $origDeepTime;

        // Benchmark OPTIMIZED implementation
        Config::$params = $config;
        $optFlatTime = $this->benchmarkLookups($flatKeys, self::N_ROUNDS);
        $optDeepTime = $this->benchmarkLookups($deepKeys, self::N_ROUNDS);
        $optTotal = $optFlatTime + $optDeepTime;

        // Output results
        $totalLookups = (count($flatKeys) + count($deepKeys)) * self::N_ROUNDS;

        echo "\n";
        echo "╔═══════════════════════════════════════════════════════════════╗\n";
        echo "║              Performance Benchmark (Before vs After)          ║\n";
        echo "╚═══════════════════════════════════════════════════════════════╝\n";
        echo "\n";
        echo "Configuration:\n";
        echo '  • Items:        '.number_format(self::N_ITEMS)."\n";
        echo '  • Rounds:       '.self::N_ROUNDS."\n";
        echo '  • Total:        '.number_format($totalLookups)." lookups\n";
        echo '  • Nesting:      '.self::DEPTH." levels\n";
        echo '  • PHP Version:  '.PHP_VERSION."\n";
        echo "\n";
        echo "┌─ ORIGINAL (array_shift) ──────────────────────────────────────┐\n";
        echo '│ Flat keys:  '.str_pad(number_format($origFlatTime, 2).' ms', 48)."│\n";
        echo '│ Deep keys:  '.str_pad(number_format($origDeepTime, 2).' ms', 48)."│\n";
        echo '│ TOTAL:      '.str_pad(number_format($origTotal, 2).' ms', 48)."│\n";
        echo "└───────────────────────────────────────────────────────────────┘\n";
        echo "\n";
        echo "┌─ OPTIMIZED (foreach + fast path) ─────────────────────────────┐\n";
        echo '│ Flat keys:  '.str_pad(number_format($optFlatTime, 2).' ms', 48)."│\n";
        echo '│ Deep keys:  '.str_pad(number_format($optDeepTime, 2).' ms', 48)."│\n";
        echo '│ TOTAL:      '.str_pad(number_format($optTotal, 2).' ms', 48)."│\n";
        echo "└───────────────────────────────────────────────────────────────┘\n";
        echo "\n";
        echo "┌─ IMPROVEMENT ─────────────────────────────────────────────────┐\n";
        echo '│ Flat keys:  '.str_pad($this->formatImprovement($origFlatTime, $optFlatTime), 48)."│\n";
        echo '│ Deep keys:  '.str_pad($this->formatImprovement($origDeepTime, $optDeepTime), 48)."│\n";
        echo '│ TOTAL:      '.str_pad($this->formatImprovement($origTotal, $optTotal), 48)."│\n";
        echo '│ Speedup:    '.str_pad(number_format($origTotal / $optTotal, 2).'x faster', 48)."│\n";
        echo "└───────────────────────────────────────────────────────────────┘\n";
        echo "\n";

        // Verify correctness with actual keys from our generated data
        $this->assertEquals(1, Config::getParam('flat1'));

        // Build the deep path dynamically based on DEPTH
        $deepPath = 'deep1';
        for ($i = 1; $i <= self::DEPTH; $i++) {
            $deepPath .= ".k{$i}";
        }
        $this->assertEquals(1, Config::getParam($deepPath));

        $this->assertEquals('default', Config::getParam('nonexistent', 'default'));
        $this->assertEquals('default', Config::getParam('deep1.k1.k2.invalid', 'default'));

        // Verify both implementations return same results
        $this->assertEquals(
            ConfigOriginal::getParam($deepPath),
            Config::getParam($deepPath),
            'Optimized and original implementations should return same results'
        );

        // Informational assertion
        $this->assertGreaterThan(0, $optTotal, 'Benchmark completed successfully');
    }

    private function formatImprovement(float $before, float $after): string
    {
        if ($before <= 0) {
            return 'n/a';
        }
        $pct = (1.0 - $after / $before) * 100.0;

        return sprintf('%+.1f%% (%s)', $pct, $pct >= 0 ? 'faster' : 'slower');
    }

    /**
     * @return array{0: array<string, mixed>, 1: array<int, string>, 2: array<int, string>}
     */
    private function generateTestData(int $n, int $depth): array
    {
        $config = [];
        $flatKeys = [];
        $deepKeys = [];

        $nFlat = (int) ($n * 0.6); // 60% flat keys
        $nDeep = $n - $nFlat;

        // Generate flat keys
        for ($i = 1; $i <= $nFlat; $i++) {
            $key = "flat{$i}";
            $config[$key] = $i;
            $flatKeys[] = $key;
        }

        // Generate deep nested keys
        for ($i = 1; $i <= $nDeep; $i++) {
            $root = "deep{$i}";
            $config[$root] = $this->buildNestedArray($i, $depth);

            // Build the full dotted path
            $path = $root;
            for ($d = 1; $d <= $depth; $d++) {
                $path .= ".k{$d}";
            }
            $deepKeys[] = $path;
        }

        return [$config, $flatKeys, $deepKeys];
    }

    /**
     * @return array<string, mixed>
     */
    private function buildNestedArray(mixed $value, int $depth): array
    {
        $node = $value;
        for ($d = $depth; $d >= 1; $d--) {
            $node = ["k{$d}" => $node];
        }

        return $node;
    }

    /**
     * @param array<int, string> $keys
     */
    private function benchmarkLookups(array $keys, int $rounds): float
    {
        $start = hrtime(true);

        for ($r = 0; $r < $rounds; $r++) {
            foreach ($keys as $key) {
                Config::getParam($key, null);
            }
        }

        $end = hrtime(true);

        return ($end - $start) / 1_000_000; // Convert to milliseconds
    }

    /**
     * @param array<int, string> $keys
     */
    private function benchmarkLookupsOriginal(array $keys, int $rounds): float
    {
        $start = hrtime(true);

        for ($r = 0; $r < $rounds; $r++) {
            foreach ($keys as $key) {
                ConfigOriginal::getParam($key, null);
            }
        }

        $end = hrtime(true);

        return ($end - $start) / 1_000_000; // Convert to milliseconds
    }
}
```

Results -
```
╔═══════════════════════════════════════════════════════════════╗
║              Performance Benchmark (Before vs After)          ║
╚═══════════════════════════════════════════════════════════════╝

Configuration:
  • Items:        10,000
  • Rounds:       100
  • Total:        1,000,000 lookups
  • Nesting:      10 levels
  • PHP Version:  8.4.5

┌─ ORIGINAL (array_shift) ──────────────────────────────────────┐
│ Flat keys:  110.99 ms                                       │
│ Deep keys:  591.17 ms                                       │
│ TOTAL:      702.15 ms                                       │
└───────────────────────────────────────────────────────────────┘

┌─ OPTIMIZED (foreach + fast path) ─────────────────────────────┐
│ Flat keys:  43.56 ms                                        │
│ Deep keys:  355.92 ms                                       │
│ TOTAL:      399.49 ms                                       │
└───────────────────────────────────────────────────────────────┘

┌─ IMPROVEMENT ─────────────────────────────────────────────────┐
│ Flat keys:  +60.7% (faster)                                 │
│ Deep keys:  +39.8% (faster)                                 │
│ TOTAL:      +43.1% (faster)                                 │
│ Speedup:    1.76x faster                                    │
└───────────────────────────────────────────────────────────────┘

Time: 00:01.116, Memory: 24.00 MB
```

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->

## Summary by CodeRabbit

- Refactor
  - Stricter typing for configuration parameters and return values.
  - More efficient retrieval of nested configuration keys while maintaining existing behavior for simple keys.
- Chores
  - Increased minimum PHP requirement to 8.1.
  - Updated development tooling (static analysis) to latest major versions.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->